backend/requirements.txt
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.2.1
docopt==0.6.2
exceptiongroup==1.0.0rc9
fastapi==0.87.0
filelock==3.9.0
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
joblib==1.2.0
mysql-connector==2.2.9
mysql-connector-repackaged==0.3.1
nltk==3.8.1
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pbr==5.11.0
pipreqs==0.4.11
platformdirs==2.6.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
redis==4.3.4
regex==2022.10.31
requests==2.28.1
rsa==4.9
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.21.0
starlette==0.21.0
stevedore==4.1.1
textblob==0.17.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
webdriver-manager==3.8.4
wrapt==1.14.1
wsproto==1.2.0
yarg==0.1.9
